requirements.txt
anytree==2.8.0
APScheduler==3.10.4
arrow==1.4.0
asgiref==3.7.2
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.5.7
cffi==1.16.0
charset-normalizer==3.3.1
colorama==0.4.6
contourpy==1.2.0
crispy-bootstrap4==2023.1
cryptography==41.0.4
cycler==0.12.1
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0
django-admin-confirm==1.0.0
django-admin-csvexport==2.2
django-admin-extra-buttons==1.5.6
django-admin-material-dashboard==1.0.10
django-adminlte3==0.1.6
django-anymail==13.0.1
django-apscheduler==0.6.2
django-cors-headers==3.14.0
django-crispy-forms==2.1
django-crontab==0.7.1
django-db-connection-pool==1.2.5
django-dbconn-retry==0.1.8
django-environ==0.10.0
django-import-export==3.2.0
django-json-widget==1.1.1
django-jsonform==2.17.0
django-jsonsuit==0.5.0
django-modeltree==0.4
django-prettyjson==0.4.1
django-storages==1.14.2
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
dnspython==2.3.0
dropbox==11.36.2
et-xmlfile==1.1.0
fastkml==1.4.0
firebase-admin==6.2.0
fonttools==4.50.0
future==0.18.3
google-api-core==2.12.0
google-api-python-client==2.104.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.2.2
google-cloud-core==2.3.3
google-cloud-firestore==2.12.0
google-cloud-storage==2.12.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpcio==1.59.0
grpcio-status==1.59.0
gunicorn==20.1.0
httplib2==0.22.0
idna==3.4
kiwisolver==1.4.5
MarkupPy==1.14
matplotlib==3.8.3
msgpack==1.0.7
numpy==1.26.4
oauthlib==3.3.1
odfpy==1.4.1
openpyxl==3.1.2
packaging==24.0
pandas==2.2.3
Pillow==9.4.0
ply==3.11
proto-plus==1.22.3
protobuf==4.24.4
psycopg2-binary==2.9.10
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pygeoif==1.6.0
PyJWT==2.8.0
pymongo==4.4.0
pyparsing==3.1.1
pyproj==3.7.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
shapely==2.1.2
six==1.16.0
SQLAlchemy==2.0.34
sqlparams==6.1.0
sqlparse==0.4.3
standardjson==0.3.1
stone==3.3.1
stripe==14.1.0
tablib==3.5.0
typing_extensions==4.8.0
tzdata==2024.2
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.0.7
whitenoise==6.4.0
xlrd==2.0.1
xlwt==1.3.0
